# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ContaNova ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /productos:
    get:
      operationId: listProductos
      tags: [productos]
      summary: Listar productos
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoria
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Producto"
    post:
      operationId: createProducto
      tags: [productos]
      summary: Crear producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoInput"
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
  /productos/{id}:
    get:
      operationId: getProducto
      tags: [productos]
      summary: Obtener producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Producto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
    patch:
      operationId: updateProducto
      tags: [productos]
      summary: Actualizar producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoUpdate"
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
    delete:
      operationId: deleteProducto
      tags: [productos]
      summary: Eliminar producto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado
...
  /facturas:
    get:
      operationId: listFacturas
      tags: [facturas]
      summary: Listar facturas de venta
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [borrador, emitida, pagada, anulada]
        - name: clienteId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de facturas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Factura"
    post:
      operationId: createFactura
      tags: [facturas]
      summary: Crear factura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacturaInput"
      responses:
        "201":
          description: Factura creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Factura"
  /facturas/{id}:
    get:
      operationId: getFactura
      tags: [facturas]
      summary: Obtener factura
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Factura con items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacturaDetalle"
    patch:
      operationId: updateFactura
      tags: [facturas]
      summary: Actualizar factura
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacturaUpdate"
      responses:
        "200":
          description: Factura actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Factura"
    delete:
      operationId: deleteFactura
      tags: [facturas]
      summary: Anular/eliminar factura
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado
...
  /cartera/abonos:
    post:
      operationId: createAbono
      tags: [cartera]
      summary: Registrar abono a factura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbonoInput"
      responses:
        "201":
          description: Abono registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Abono"

  # ─── PLAN DE CUENTAS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Producto:
      type: object
      required:
        - id
        - codigo
        - nombre
        - precioVenta
        - stock
        - activo
        - createdAt
      properties:
        id:
          type: integer
        codigo:
          type: string
        nombre:
          type: string
        descripcion:
          type: ["string", "null"]
        categoria:
          type: ["string", "null"]
        unidad:
          type: string
        precioVenta:
          type: number
        precioCosto:
          type: ["number", "null"]
        stock:
          type: number
        stockMinimo:
          type: ["number", "null"]
        activo:
          type: boolean
        createdAt:
          type: string
    ProductoInput:
      type: object
      required: [codigo, nombre, precioVenta, unidad]
      properties:
        codigo:
          type: string
          minLength: 1
        nombre:
          type: string
          minLength: 1
        descripcion:
          type: string
        categoria:
          type: string
        unidad:
          type: string
        precioVenta:
          type: number
        precioCosto:
          type: number
        stock:
          type: number
        stockMinimo:
          type: number
    ProductoUpdate:
      type: object
      properties:
        codigo:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        categoria:
          type: string
        unidad:
          type: string
        precioVenta:
          type: number
        precioCosto:
          type: number
        stock:
          type: number
        stockMinimo:
          type: number
        activo:
          type: boolean
...
    FacturaItem:
      type: object
      required: [id, productoId, descripcion, cantidad, precioUnitario, subtotal]
      properties:
        id:
          type: integer
        productoId:
          type: ["integer", "null"]
        descripcion:
          type: string
        cantidad:
          type: number
        precioUnitario:
          type: number
        descuento:
          type: ["number", "null"]
        subtotal:
          type: number
    FacturaItemInput:
      type: object
      required: [descripcion, cantidad, precioUnitario]
      properties:
        productoId:
          type: integer
        descripcion:
          type: string
        cantidad:
          type: number
        precioUnitario:
          type: number
        descuento:
          type: number
    Factura:
      type: object
      required:
        - id
        - numero
        - clienteId
        - clienteNombre
        - fecha
        - fechaVencimiento
        - subtotal
        - impuesto
        - total
        - estado
        - createdAt
      properties:
        id:
          type: integer
        numero:
          type: string
        clienteId:
          type: integer
        clienteNombre:
          type: string
        fecha:
          type: string
        fechaVencimiento:
          type: string
        subtotal:
          type: number
        descuento:
          type: ["number", "null"]
        impuesto:
          type: number
        total:
          type: number
        saldoPendiente:
          type: ["number", "null"]
        estado:
          type: string
          enum: [borrador, emitida, pagada, anulada]
        notas:
          type: ["string", "null"]
        createdAt:
          type: string
    FacturaDetalle:
      type: object
      required:
        - id
        - numero
        - clienteId
        - clienteNombre
        - fecha
        - fechaVencimiento
        - subtotal
        - impuesto
        - total
        - estado
        - items
        - createdAt
      properties:
        id:
          type: integer
        numero:
          type: string
        clienteId:
          type: integer
        clienteNombre:
          type: string
        fecha:
          type: string
        fechaVencimiento:
          type: string
        subtotal:
          type: number
        descuento:
          type: ["number", "null"]
        impuesto:
          type: number
        total:
          type: number
        saldoPendiente:
          type: ["number", "null"]
        estado:
          type: string
          enum: [borrador, emitida, pagada, anulada]
        notas:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/FacturaItem"
        createdAt:
          type: string
    FacturaInput:
      type: object
      required: [clienteId, fecha, fechaVencimiento, items]
      properties:
        clienteId:
          type: integer
        fecha:
          type: string
          format: date
        fechaVencimiento:
          type: string
          format: date
        descuento:
          type: number
        notas:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/FacturaItemInput"
    FacturaUpdate:
      type: object
      properties:
        fecha:
          type: string
        fechaVencimiento:
          type: string
        estado:
          type: string
          enum: [borrador, emitida, pagada, anulada]
        notas:
          type: string
...
    AbonoInput:
      type: object
      required: [facturaId, monto, fecha]
      properties:
        facturaId:
          type: integer
        monto:
          type: number
        fecha:
          type: string
          format: date
        descripcion:
          type: string
    Abono:
      type: object
      required: [id, facturaId, monto, fecha, createdAt]
      properties:
        id:
          type: integer
        facturaId:
          type: integer
        monto:
          type: number
        fecha:
          type: string
        descripcion:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── CUENTAS CONTABLES ────────────────────────────────────────────────────
...