# lib/api-zod/src/generated/types/facturaDetalle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ContaNova ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { FacturaDetalleEstado } from './facturaDetalleEstado';
import type { FacturaItem } from './facturaItem';

export interface FacturaDetalle {
  id: number;
  numero: string;
  clienteId: number;
  clienteNombre: string;
  fecha: string;
  fechaVencimiento: string;
  subtotal: number;
  /** @nullable */
  descuento?: number | null;
  impuesto: number;
  total: number;
  /** @nullable */
  saldoPendiente?: number | null;
  estado: FacturaDetalleEstado;
  /** @nullable */
  notas?: string | null;
  items: FacturaItem[];
  createdAt: string;
}